- add new projects to this version - run mvn eclipse:eclipse again and import plexus-remote and theme-wwu.
 - this commit will revert revision 3604, we will implement a better solution of naming discussion then implemented in this version.


git-svn-id: https://svn.code.sf.net/p/openuss/code/branches/openuss-plexus-usability@3605 60046477-0c10-0410-bf4b-e552a4ff1768

+++ plexus/plexus-core/src/main/java/org/openuss/lecture/CourseServiceImpl.java
@@ -66,7 +66,7 @@ public class CourseServiceImpl extends org.openuss.lecture.CourseServiceBase {
 		
 		
 		// Create default Group for Course
-		Group participantsGroup = getSecurityService().createGroup("COURSE_" + courseEntity.getId() + "_PARTICIPANTS", "autogroup_administrator_label", "", GroupType.ADMINISTRATOR);
+		Group participantsGroup = getSecurityService().createGroup("COURSE_" + courseEntity.getId() + "_PARTICIPANTS", "autogroup_participant_label", null, GroupType.PARTICIPANT);
 		Set<Group> groups = courseEntity.getGroups();
 		if (groups == null){
 			groups = new HashSet<Group>();
@@ -88,7 +88,9 @@ public class CourseServiceImpl extends org.openuss.lecture.CourseServiceBase {
 	protected void handleRemoveCourse(Long courseId) throws Exception {
 		Validate.notNull(courseId, "CourseId cannot be null.");
 		Course course = (Course) this.getCourseDao().load(courseId);
-		Validate.notNull(course, "No course entity found with the corresponding courseId " + courseId);
+		if (course == null) {
+			throw new CourseServiceException("No course entity found with the corresponding courseId "+ courseId);
+		}
 
 		// Remove Security
 		this.getSecurityService().removeAllPermissions(course);
@@ -115,7 +117,7 @@ public class CourseServiceImpl extends org.openuss.lecture.CourseServiceBase {
 
 	protected CourseInfo handleFindCourse(Long courseId) {
 		Validate.notNull(courseId, "CourseId cannot be null.");
-		return this.getCourseDao().toCourseInfo(this.getCourseDao().load(courseId));
+		return (CourseInfo) this.getCourseDao().load(CourseDao.TRANSFORM_COURSEINFO, courseId); 
 	}
 
 	@Override
@@ -589,5 +591,4 @@ public class CourseServiceImpl extends org.openuss.lecture.CourseServiceBase {
 			return self.equals(found);
 		}
 	}
-
 }
